+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - [#1209](https://github.com/openscope/openscope/issues/1209) - Update TUI Airways callsign
 - [#604](https://github.com/openscope/openscope/issues/604) - Adds support for dataBlock timeshare data
 - [#678](https://github.com/openscope/openscope/issues/678) - Removes wind vane visualization from `CanvasController`
+- [#1238](https://github.com/openscope/openscope/issues/1238) - Route and reroute should include IFR clearance